+++ src/pybullet_planning/interfaces/robots/__init__.py
@@ -14,8 +14,6 @@ Body
     :toctree: generated/
     :nosignatures:
 
-    vertices_from_rigid
-
 Collision checking
 --------------------